bring pages that ping slackbot into the docs using the mdx idea that lets you import react modules into markup https://docusaurus.io/docs/markdown-features/react
