+++ docs/Learn/get-invite-key.md
@@ -1,12 +1,8 @@
 ---
 sidebar_position: 4
 ---
+import GetInviteKey from '@site/src/components/pages/get-invite-key';
+
 # Get an invite
 
-### Request invite
-Would you like to use https://estuary.tech to make Filecoin storage deals?
-
-[Please fill out the form here](/get-invite-key)! 
-
-We'll get back to you either over Twitter or e-mail if we think you have a meaningful public data set.
-
+<GetInviteKey />
