+++ docs/Learn/get-provider-added-cn.md
@@ -1,13 +1,9 @@
 ---
 sidebar_position: 6
 ---
+
+import GetProviderAdded from '@site/src/components/pages/get-provider-added-cn';
+
 # 注册服务商
 
-### Storage provider application
-
-想要将你的存储供应商ID加入到 https://estuary.tech 并且收到已验证存储订单? 
-
-[请填写申请表，](/get-provider-added-cn)
-
-如果我们相信你的存储系统是有资格的我们会把它加入Estuary的存储服务商列表中，
-加入之后你应该就会开始收到存储订单。
+<GetProviderAdded />
