+++ docs/Learn/get-provider-added.md
@@ -1,13 +1,8 @@
 ---
 sidebar_position: 5
 ---
+import GetProviderAdded from '@site/src/components/pages/get-provider-added';
+
 # Storage Provider Application
 
-
-### Storage Provider Application
-Would you like to get your provider added to https://estuary.tech and receive verified Filecoin storage deals? 
-
-[Please fill out this form!](/get-provider-added)
-
-If we believe your provider is qualified we will add it to our provider index and you will receive storage deals.
-
+<GetProviderAdded />
