+++ docs/give-feedback.md
@@ -1,5 +1,8 @@
+import Feedback from '@site/src/components/pages/feedback';
+
 # Feedback
 
-Would you like to provide feedback for Estuary? [Please submit it here.](/feedback)
+
+<Feedback />
