+++ legacy/pages/Page.module.scss
@@ -17,3 +17,8 @@
   display: flex;
   align-items: flex-start;
 }
+
+
+.ul {
+  margin-top: 12px;
+}
